refactor: split gym.rs into sub-modules (types, scenarios, executor, reporting)
Split the 1,492-line src/gym.rs into src/gym/ with five sub-modules,
each under the 400 impl-line limit:

- types.rs (191 lines): public Benchmark* structs/enums and Display impls
- scenarios.rs (319 lines): BENCHMARK_SCENARIOS const, scenario lookup,
  and class-specific checks
- executor.rs (390 impl lines): execute_scenario, metric types/derivation,
  and 2 existing tests
- reporting.rs (397 lines): text rendering, comparison logic, IO helpers,
  and stored report types
- mod.rs (216 lines): public API (run_benchmark_scenario, run_benchmark_suite,
  compare_latest_benchmark_runs), runtime composition helpers, and re-exports

All 656 lib tests pass. All public API paths via crate::gym::* preserved.
