+++ src/gym/scenarios.rs
@@ -1,0 +1,319 @@
+use crate::error::{SimardError, SimardResult};
+use crate::handoff::RuntimeHandoffSnapshot;
+use crate::runtime::RuntimeTopology;
+
+use super::types::{BenchmarkCheckResult, BenchmarkClass, BenchmarkScenario};
+
+const BENCHMARK_SCENARIOS: [BenchmarkScenario; 9] = [
+    BenchmarkScenario {
+        id: "repo-exploration-local",
+        title: "Repo exploration on local harness",
+        description: "Exercise a bounded repo-exploration task through the gym identity on the single-process local harness.",
+        class: BenchmarkClass::RepoExploration,
+        identity: "simard-gym",
+        base_type: "local-harness",
+        topology: RuntimeTopology::SingleProcess,
+        objective: "Inspect repository structure, identify likely extension points, and summarize where benchmark and runtime changes should land.",
+        expected_min_runtime_evidence: 3,
+    },
+    BenchmarkScenario {
+        id: "docs-refresh-copilot",
+        title: "Documentation refresh through copilot-sdk alias",
+        description: "Exercise a documentation-oriented benchmark while preserving the explicit copilot-sdk selection and honest local-harness implementation identity.",
+        class: BenchmarkClass::Documentation,
+        identity: "simard-gym",
+        base_type: "copilot-sdk",
+        topology: RuntimeTopology::SingleProcess,
+        objective: "Produce a concise documentation-oriented execution summary for the current repository state and report the relevant reflected runtime contracts.",
+        expected_min_runtime_evidence: 3,
+    },
+    BenchmarkScenario {
+        id: "safe-code-change-rusty-clawd",
+        title: "Safe code change style task on rusty-clawd",
+        description: "Exercise a bounded safe-change objective on the distinct rusty-clawd backend through the loopback multi-process topology.",
+        class: BenchmarkClass::SafeCodeChange,
+        identity: "simard-gym",
+        base_type: "rusty-clawd",
+        topology: RuntimeTopology::MultiProcess,
+        objective: "Plan a narrow, reviewable runtime change and summarize the exact evidence an operator would inspect before approving it.",
+        expected_min_runtime_evidence: 4,
+    },
+    BenchmarkScenario {
+        id: "composite-session-review",
+        title: "Composite identity session quality review",
+        description: "Exercise the composite engineer identity as a session-quality benchmark so the starter suite covers the shipped composite identity as well as the dedicated gym identity.",
+        class: BenchmarkClass::SessionQuality,
+        identity: "simard-composite-engineer",
+        base_type: "local-harness",
+        topology: RuntimeTopology::SingleProcess,
+        objective: "Run a disciplined bounded engineering session, preserve evidence, and produce a concise operator-facing summary of what happened.",
+        expected_min_runtime_evidence: 3,
+    },
+    BenchmarkScenario {
+        id: "interactive-terminal-driving",
+        title: "Interactive terminal driving on terminal-shell",
+        description: "Exercise the engineer identity through the terminal-shell base type by launching a bounded interactive child process, waiting for prompts, and sending follow-up inputs like an operator validating generic PTY-driven control flow.",
+        class: BenchmarkClass::SessionQuality,
+        identity: "simard-engineer",
+        base_type: "terminal-shell",
+        topology: RuntimeTopology::SingleProcess,
+        objective: "working-directory: .\ncommand: sh -c 'printf \"terminal-ready\\n\"; while IFS= read -r line; do if [ \"$line\" = \"ack\" ]; then printf \"terminal-ack\\n\"; elif [ \"$line\" = \"exit\" ]; then printf \"terminal-bye\\n\"; break; else printf \"echo:%s\\n\" \"$line\"; fi; done'\nwait-for: terminal-ready\ninput: ack\nwait-for: terminal-ack\ninput: exit\nwait-for: terminal-bye",
+        expected_min_runtime_evidence: 6,
+    },
+    // --- Benchmark scenarios: real gym challenges ---
+    BenchmarkScenario {
+        id: "repo-exploration-deep-scan",
+        title: "Deep repository structure analysis",
+        description: "Explore the Simard repository itself to identify project structure, key source files, dependency graph from Cargo.toml, and executable entry points. Scored on completeness of structural discovery.",
+        class: BenchmarkClass::RepoExploration,
+        identity: "simard-gym",
+        base_type: "local-harness",
+        topology: RuntimeTopology::SingleProcess,
+        objective: "Analyze the repository at /home/azureuser/src/Simard. Identify: (1) the top-level project structure and key directories, (2) all Cargo.toml dependencies and their purposes, (3) the main entry point(s) in src/main.rs or src/lib.rs, (4) at least five public modules and their responsibilities. Produce a structured summary covering all four areas.",
+        expected_min_runtime_evidence: 4,
+    },
+    BenchmarkScenario {
+        id: "doc-generation-public-fn",
+        title: "Generate doc comment for public function",
+        description: "Given a source file in the Simard repository, generate a syntactically valid Rust doc comment for a public function. Scored on whether the comment is valid rustdoc, mentions parameters and return type, and accurately describes behavior.",
+        class: BenchmarkClass::Documentation,
+        identity: "simard-gym",
+        base_type: "local-harness",
+        topology: RuntimeTopology::SingleProcess,
+        objective: "Read the function `pub fn benchmark_scenarios()` in src/gym.rs. Generate a complete Rust doc comment (/// style) for it that: (1) describes what the function returns, (2) mentions the BenchmarkScenario type, (3) notes the static lifetime of the returned slice, (4) is syntactically valid rustdoc. Output the doc comment text.",
+        expected_min_runtime_evidence: 3,
+    },
+    BenchmarkScenario {
+        id: "safe-code-change-add-derive",
+        title: "Add Debug derive to a struct",
+        description: "Execute a safe, minimal code change: add a #[derive(Debug)] attribute to a struct that lacks it. Scored on whether the change compiles successfully and the derive is correctly placed.",
+        class: BenchmarkClass::SafeCodeChange,
+        identity: "simard-gym",
+        base_type: "local-harness",
+        topology: RuntimeTopology::SingleProcess,
+        objective: "Identify a public struct in the Simard codebase that does not derive Debug. Plan a minimal change to add #[derive(Debug)] to it. Describe the exact file, line, and change required. Verify the change would compile by checking for any generic parameter constraints that might prevent Debug derivation.",
+        expected_min_runtime_evidence: 3,
+    },
+    BenchmarkScenario {
+        id: "test-writing-unit-case",
+        title: "Write a unit test for a public function",
+        description: "Given a function signature, write a unit test that exercises its basic behavior. Scored on whether the test compiles, covers the primary code path, and uses appropriate assertions.",
+        class: BenchmarkClass::TestWriting,
+        identity: "simard-gym",
+        base_type: "local-harness",
+        topology: RuntimeTopology::SingleProcess,
+        objective: "Write a unit test for the function `pub fn goal_slug(title: &str) -> String` defined in src/goals.rs. The test should: (1) call goal_slug with a representative input string containing uppercase letters, spaces, and special characters, (2) assert the output matches expected slug format (lowercase, hyphen-separated, no leading/trailing hyphens), (3) be a valid #[test] function that compiles and runs.",
+        expected_min_runtime_evidence: 3,
+    },
+];
+
+pub fn benchmark_scenarios() -> &'static [BenchmarkScenario] {
+    &BENCHMARK_SCENARIOS
+}
+
+pub(super) fn resolve_benchmark_scenario(scenario_id: &str) -> SimardResult<BenchmarkScenario> {
+    benchmark_scenarios()
+        .iter()
+        .copied()
+        .find(|candidate| candidate.id == scenario_id)
+        .ok_or_else(|| SimardError::BenchmarkScenarioNotFound {
+            scenario_id: scenario_id.to_string(),
+        })
+}
+
+/// Produce additional scenario-class-specific checks based on the scenario type.
+///
+/// These supplement the generic 8-check baseline with checks tailored to each
+/// `BenchmarkClass`: structural discovery for repo exploration, doc validity
+/// for documentation, compilation evidence for safe code changes, and test
+/// structure for test writing scenarios.
+pub(super) fn class_specific_checks(
+    scenario: &BenchmarkScenario,
+    outcome: &crate::runtime::SessionOutcome,
+    exported: &RuntimeHandoffSnapshot,
+) -> Vec<BenchmarkCheckResult> {
+    let summary = outcome.execution_summary.to_lowercase();
+    let plan = outcome.plan.to_lowercase();
+    let reflection = outcome.reflection.summary.to_lowercase();
+    let combined = format!("{summary} {plan} {reflection}");
+
+    match scenario.class {
+        BenchmarkClass::RepoExploration => {
+            let structure_mentioned = combined.contains("src/")
+                || combined.contains("directory")
+                || combined.contains("structure")
+                || combined.contains("module");
+            let deps_mentioned = combined.contains("cargo.toml")
+                || combined.contains("dependenc")
+                || combined.contains("crate");
+            let entry_mentioned = combined.contains("main.rs")
+                || combined.contains("lib.rs")
+                || combined.contains("entry point")
+                || combined.contains("entry-point");
+            vec![
+                BenchmarkCheckResult {
+                    id: "repo-structure-discovered".to_string(),
+                    passed: structure_mentioned,
+                    detail: format!(
+                        "execution output {} project structure references",
+                        if structure_mentioned {
+                            "contains"
+                        } else {
+                            "lacks"
+                        }
+                    ),
+                },
+                BenchmarkCheckResult {
+                    id: "repo-dependencies-identified".to_string(),
+                    passed: deps_mentioned,
+                    detail: format!(
+                        "execution output {} dependency references",
+                        if deps_mentioned { "contains" } else { "lacks" }
+                    ),
+                },
+                BenchmarkCheckResult {
+                    id: "repo-entry-points-found".to_string(),
+                    passed: entry_mentioned,
+                    detail: format!(
+                        "execution output {} entry point references",
+                        if entry_mentioned { "contains" } else { "lacks" }
+                    ),
+                },
+            ]
+        }
+        BenchmarkClass::Documentation => {
+            let has_doc_syntax = combined.contains("///")
+                || combined.contains("doc comment")
+                || combined.contains("rustdoc")
+                || combined.contains("documentation");
+            let mentions_params = combined.contains("param")
+                || combined.contains("argument")
+                || combined.contains("return")
+                || combined.contains("-> ");
+            vec![
+                BenchmarkCheckResult {
+                    id: "doc-comment-syntax-valid".to_string(),
+                    passed: has_doc_syntax,
+                    detail: format!(
+                        "execution output {} doc comment syntax",
+                        if has_doc_syntax {
+                            "references"
+                        } else {
+                            "lacks"
+                        }
+                    ),
+                },
+                BenchmarkCheckResult {
+                    id: "doc-params-return-covered".to_string(),
+                    passed: mentions_params,
+                    detail: format!(
+                        "execution output {} parameter/return documentation",
+                        if mentions_params { "includes" } else { "lacks" }
+                    ),
+                },
+            ]
+        }
+        BenchmarkClass::SafeCodeChange => {
+            let compilation_evidence = combined.contains("compil")
+                || combined.contains("cargo build")
+                || combined.contains("cargo check")
+                || combined.contains("build succeed")
+                || combined.contains("no errors");
+            let change_described = combined.contains("derive")
+                || combined.contains("change")
+                || combined.contains("modif")
+                || combined.contains("diff");
+            vec![
+                BenchmarkCheckResult {
+                    id: "code-change-compilation-checked".to_string(),
+                    passed: compilation_evidence,
+                    detail: format!(
+                        "execution output {} compilation verification",
+                        if compilation_evidence {
+                            "includes"
+                        } else {
+                            "lacks"
+                        }
+                    ),
+                },
+                BenchmarkCheckResult {
+                    id: "code-change-described".to_string(),
+                    passed: change_described,
+                    detail: format!(
+                        "execution output {} change description",
+                        if change_described {
+                            "includes"
+                        } else {
+                            "lacks"
+                        }
+                    ),
+                },
+            ]
+        }
+        BenchmarkClass::TestWriting => {
+            let has_test_annotation = combined.contains("#[test]")
+                || combined.contains("test function")
+                || combined.contains("unit test");
+            let has_assertion = combined.contains("assert")
+                || combined.contains("expect")
+                || combined.contains("should_eq")
+                || combined.contains("assert_eq");
+            let covers_basic_case = combined.contains("input")
+                || combined.contains("call")
+                || combined.contains("invoke")
+                || combined.contains("result");
+            vec![
+                BenchmarkCheckResult {
+                    id: "test-structure-valid".to_string(),
+                    passed: has_test_annotation,
+                    detail: format!(
+                        "execution output {} test annotation/structure",
+                        if has_test_annotation {
+                            "includes"
+                        } else {
+                            "lacks"
+                        }
+                    ),
+                },
+                BenchmarkCheckResult {
+                    id: "test-has-assertions".to_string(),
+                    passed: has_assertion,
+                    detail: format!(
+                        "execution output {} assertions",
+                        if has_assertion { "includes" } else { "lacks" }
+                    ),
+                },
+                BenchmarkCheckResult {
+                    id: "test-covers-basic-case".to_string(),
+                    passed: covers_basic_case,
+                    detail: format!(
+                        "execution output {} basic case coverage",
+                        if covers_basic_case {
+                            "includes"
+                        } else {
+                            "lacks"
+                        }
+                    ),
+                },
+            ]
+        }
+        BenchmarkClass::SessionQuality => {
+            // Session quality scenarios rely on the generic checks.
+            let session_summary_present =
+                !outcome.execution_summary.trim().is_empty() && exported.memory_records.len() >= 2;
+            vec![BenchmarkCheckResult {
+                id: "session-quality-summary-adequate".to_string(),
+                passed: session_summary_present,
+                detail: format!(
+                    "session produced {} memory records with {} execution summary",
+                    exported.memory_records.len(),
+                    if outcome.execution_summary.trim().is_empty() {
+                        "empty"
+                    } else {
+                        "non-empty"
+                    }
+                ),
+            }]
+        }
+    }
+}
